moved `entities` and `base` into own domain
