+++ MalawiMeta.Api/Domain/Entities/District.cs
@@ -1,6 +1,4 @@
-using MalawiMeta.Api.Entities.Base;
-
-namespace MalawiMeta.Api.Entities;
+namespace MalawiMeta.Api.Domain.Entities;
 
 public enum Region
 {
